requirements.txt
Flask>=0.11
Flask-Script>=2.0.5
Flask-PyMongo>=2.3.0
pymongo>=3.4.0
py-bcrypt>=0.4
jsonschema>=2.5.1